Rename src/util/github.ts to fix filename casing issue, next commit will fix for good
